requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
cmudict==1.0.13
cymem==2.0.7
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Distance==0.1.3
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.2.0/en_core_web_md-3.2.0-py3-none-any.whl
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
g2p-en==2.1.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.2
inflect==6.0.2
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
langcodes==3.3.0
lemminflect==0.2.3
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.8
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
preshed==3.0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.8.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-datamuse==1.3.2
python-dateutil==2.8.2
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
Send2Trash==1.8.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
smart-open==6.3.0
sniffio==1.3.0
SoundsLike==0.0.11
soupsieve==2.3.2.post1
spacy==3.2.4
spacy-legacy==3.0.10
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
terminado==0.17.0
thinc==8.0.17
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.13
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
zipp==3.11.0
